requirements.txt
pytz
numpy
Markdown
langchain_core
pyautogui
Pillow
scikit-image
PyQt6
openai
tiktoken
google-genai
dotenv
demjson3
pathvalidate
docstring_parser
sqlalchemy
pydantic-settings
jinja2
humps
composio
colorama
anthropic
httpx_sse
Markdown
rapidfuzz
llama_index